all files moved to /lib
