+++ src/frontend/hosting/src/lib/store/auth_store.ts
